token usage in separated file
